feat(testlib): refactor setupFiles to avoid duplication
