+++ .changeset/late-baboons-love.md
@@ -1,0 +1,5 @@
+---
+"@smartthings/cli-testlib": patch
+---
+
+refactor setupFiles to avoid duplication and remove setup function in favor of self executing module
