+++ packages/cli/src/__tests__/commands/apps/create.test.ts
@@ -5,15 +5,6 @@ import { tableFieldDefinitions } from '../../../lib/commands/apps/apps-util'
 import { addPermission } from '../../../lib/aws-utils'
 
 
-jest.mock('@smartthings/cli-lib', () => {
-	const originalLib = jest.requireActual('@smartthings/cli-lib')
-
-	return {
-		...originalLib,
-		inputAndOutputItem: jest.fn(),
-	}
-})
-
 jest.mock('../../../lib/aws-utils')
 
 describe('AppCreateCommand', () => {
