+++ packages/cli/src/__tests__/commands/apps/settings.test.ts
@@ -4,15 +4,6 @@ import AppSettingsCommand from '../../../commands/apps/settings'
 import { buildTableOutput, chooseApp } from '../../../lib/commands/apps/apps-util'
 
 
-jest.mock('@smartthings/cli-lib', () => {
-	const originalLib = jest.requireActual('@smartthings/cli-lib')
-
-	return {
-		...originalLib,
-		outputItem: jest.fn(),
-	}
-})
-
 jest.mock('../../../lib/commands/apps/apps-util')
 
 describe('AppSettingsCommand', () => {
