+++ packages/cli/src/__tests__/commands/locations/update.test.ts
@@ -4,15 +4,6 @@ import { inputAndOutputItem } from '@smartthings/cli-lib'
 import { chooseLocation } from '../../../commands/locations'
 
 
-jest.mock('@smartthings/cli-lib', () => {
-	const originalLib = jest.requireActual('@smartthings/cli-lib')
-
-	return {
-		...originalLib,
-		inputAndOutputItem: jest.fn(),
-	}
-})
-
 jest.mock('../../../commands/locations')
 
 describe('LocationsUpdateCommand', () => {
