--- packages/cli/src/__tests__/setup/jest.setup.ts
@@ -1,4 +1,0 @@
-import { setup } from '@smartthings/cli-testlib'
-
-
-setup()
