--- packages/testlib/jest.setup.ts.example
@@ -1,4 +1,0 @@
-import { setup } from '@smartthings/cli-testlib'
-
-
-setup()
